refactored delete user button
